requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
reportlab>=4.0.0
openai>=1.0.0
